Time: 231 ms (9.89%) | Memory: 108 MB (5.28%) - LeetSync

+++ 1966-frequency-of-the-most-frequent-element/frequency-of-the-most-frequent-element.cpp
@@ -1,21 +1,21 @@
 class Solution {
 
-    bool possible(vector<int>&nums,vector<long long>&prefix,int n,int k)
+    bool possible(vector<int>&nums,vector<long>&prefix,long n,int k)
     {
-       long long lhs = (long long)((long long)nums[n-1]*(long long)n);
-       long long rhs = prefix[n-1] + (long long)k;
+        long lhs = (long)((long)nums[n-1]*n);
+        long rhs = prefix[n-1] + k;
 
-       if(lhs <= rhs)return true;
+        if(lhs <= rhs)return true;
 
-       for(int i = n ; i < nums.size() ; i++)
-       {
-           lhs = (long long)((long long)nums[i]*(long long)n);
-           rhs = (prefix[i]-prefix[i-n]) + (long long)k;
+        for(int i = n ; i < nums.size() ; i++)
+        {
+            lhs = (long)((long)nums[i]*n);
+            rhs = prefix[i] - prefix[i-n] + k;
 
-           if(lhs <= rhs)return true;
-       }
+            if(lhs <= rhs)return true;
+        }
 
-       return false;
+        return false;
     }
 
 public:
@@ -23,7 +23,7 @@ public:
     {
         sort(nums.begin(),nums.end());
 
-        vector<long long>prefix(nums.size());
+        vector<long>prefix(nums.size());
         prefix[0] = nums[0];
 
         for(int i = 1 ; i < nums.size() ; i++)
@@ -31,12 +31,12 @@ public:
             prefix[i] = prefix[i-1] + nums[i];
         }    
 
-        int lo = 1;
-        int hi = nums.size();
+        long lo = 1;
+        long hi = nums.size();
 
         while(hi - lo > 1)
         {
-            int mid = (lo + hi)>>1;
+            long mid = (lo + hi)>>1;
 
             if(possible(nums,prefix,mid,k))
             {
